added functions listeners
