+++ src/filtering-waypoint.js
@@ -1,10 +1,8 @@
-import renderWaypoint from "./create-waypoint";
 
 const filterList = document.querySelector(`.trip-filter`);
 
 const bindEventRender = () => {
   filterList.addEventListener(`change`, () => {
-    renderWaypoint();
   });
 };
